Prevent multiple runtime config releases from ending up in the deployment
Releases from runtime configs were ending up in the deployment, even if the addon
from that runtime config was actually excluded from the deployment based
on include/exclude placement rules. This would happen when there was a different
runtime config that included the same release, but a different version.

For example, if I have two different runtime configs that both include the "syslog"
release, but one includes version "1" and the other includes version "2".
Both of these runtime configs have "include" placement rules that limit which
"deployments" they will apply to, for example "deployment_with_syslog_v1" and
"deployment_with_syslog_v2".

Trying to deploy "deployment_with_syslog_v1" would lead to both runtime configs
being merged together in the code. So now we have both releases "1" and "2" of
syslog release in the merged runtime config. We later remove the "addons" that do
not apply to the deployment, and include only "releases" that apply to those "addons".

But since the "addons" only specify 'release: "syslog"', it appears that both
release versions should actually be included. You then end up getting this error:

Runtime manifest specifies release 'syslog' with version as '1'.
This conflicts with version '2' specified in the deployment manifest.

Which is not 100% accurate, and shouldn't be happening.

This commit changes it so we only add the "releases" from a runtime config
to the deployment once we've verified the "addon" applies to the deployment.

[#179859886]

Signed-off-by: Joseph Palermo <jpalermo@pivotal.io>