`test-integration.sh` is idempotent
One of the challenges facing developers is that the BOSH integration
tests, until now, weren't idempotent--a developer couldn't hijack a
failed container, test their fix, and re-run integration test. In the
worst case, this can lead to a series of "I hope this fixes the failure"
commits.

This commit attempts to make the `test-integration.sh` idempotent, with
special thanks to Long Nguyen who blazed a trail.

We are confident that the MySQL tests are idempotent, diffident that the
PostgreSQL tests are.