Fix test in external_cpi_spec.rb
This test was failing on our CI because it expects ruby to be present
in the env that is passed to Open3.popen3 in external_cpi.rb. On our CI,
ruby lives in /usr/local/bin. We've modified the test to add this path.
We've also fixed some typos + used a single before block as a stylistic
preference.

Signed-off-by: Rajan Agaskar <ragaskar@vmware.com>