Aborts errand run if instance(s) are stopped
A customer discovered a problem when running an errand on a VM that
was stopped with `bosh stop`. It is our understanding that Bosh will
never `stop` a VM automatically; this is a manual process that an
operator must perform.

Because we can't know why an operator stopped an instance that might
be used to run an errand, it's safer to just halt the planning process
and raise an error that informs the operator that manual intervention
is required.

[#179289747] run-errand behavior is odd if `bosh stop` is run on the errand VM

Signed-off-by: Kenneth Lakin <klakin@vmware.com>