Revert behavior change to only export first level dependent packages with bosh export-release
Bosh tracks package dependencies with a dependency_key field. We are unable to calculate this
if we don't have all the dependent packages. The tracked attributes of a package, such as
fingerprint, version and sha1, do NOT track changes in dependent packages.

So if a dependent package changes, we have no way to differentiate a new version of the
main package from an older version.

Signed-off-by: Joseph Palermo <jpalermo@pivotal.io>