Adds .gitkeep to keep the extensions dir around
The release creation process requires the `vendor/cache/extensions`
directory to be present. However, the typical contents of that directory
are OS and/or CPU architecture-dependent, and probably should never be
commited to source control. Recent commits removed that directory, which
caused our release build process to fail.

So, this commit adds a `.gitkeep` file into `vendor/cache/extensions`,
which appears to satisfy the release creation process.

Another way to have resolved this issue would have been to remove the
`vendor/cache/extensions/**` pattern from `packages/director/spec` that
was introduced in commit `0f60a4c`. Because there's no readily available
documentation that describes what would go into this directory, we're
erring on the side of caution.

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>