Increase Key Size for generated test certs
on linux when you try running integration tests,
almost every test fails

with:

```sh

```

but that is a red herring. what actually happens is:

```sh
cat tmp/integration-tests-workspace/pid-25162/sandbox/logs/error.log
2022/03/04 16:38:23 [warn] 25184#0: the "ssl" directive is deprecated, use the "listen ... ssl" directive instead in /home/kkiess/workspace/bosh/src/tmp/integration-tests-workspace/pid-25162/sandbox/nginx.conf:27
2022/03/04 16:38:23 [emerg] 25184#0: SSL_CTX_use_certificate("/home/kkiess/workspace/bosh/src/bosh-dev/assets/sandbox/ca/certs/server.crt") failed (SSL: error:140AB18F:SSL routines:SSL_CTX_use_certificate:ee key too small)
```

the nginx is not starting because the key size is too small.

this upptes the script to generate 2048 byte keys and also
regenerates the current certs.