Specify `verify_peer` for em-http-request
The version of em-http-request emits a warning if you create an
EM::HttpRequest object without specifying the `verify_peer` option. This
was because the previously default was considered unsafe per
CVE-2020-13482.

In order to use `verify_peer: true` would require a larger effort to
pass and trust the certificates between jobs. This would be non-trivial,
breaking, and only affects local communication.

[#181347667] Get rid of warnings from em-http-request

Signed-off-by: Long Nguyen <nguyenlo@vmware.com>