Accommodate M1 macs: bump gems sys-filesystem, ffi
This fixes the following error in the ffi gem when running, among other
things, the BOSH Director unit specs on an M1 mac:

```
FFI::NotFoundError:
  Function 'getmntinfo64' not found in [libc.dylib]
```

The ffi gem depends on the sys-filesystem gem which, prior to version
[1.4.1](https://github.com/djberg96/sys-filesystem/issues/48#issuecomment-752408506),
used the deprecated `getmntinfo64()` library call. From the `getmntinfo`
man page:

> The getmntinfo64() routine is equivalent to its corresponding
non-64-suffixed routine, when 64-bit inodes are in effect.  It was added
before there was support for the symbol variants, and _so is now
deprecated_.
before including header files to force 64-bit inode support. [italics
mine]

This commit bumps the sys-filesystem gem in both the Gemfile and the
BOSH Director gemspec. We bump it directly in the gemspec, and bump it
indirectly in the Gemfile by bumping the ffi gem, which depends on the
newer sys-filesystem gem.