Reset unresponsive VM count for deleted deployment
We received report of a bug where the Prometheus metrics server
continued to report unresponsive VMs within a specific BOSH deployment
after that deployment had been deleted. This is because the metrics rely
on the Health Monitor `/unresponsive_agents` API endpoint, which only
reports statistics for current deployments. After the deletion, the
Health Monitor endpoint stopped reporting the deployment, so the
Prometheus updater kept the last reported metrics.

The Prometheus client gem does not provide a way to delete metrics
entirely; therefore, we set the unresponsive agent count to 0 instead.

[#182015765] Prometheus metrics are not cleaned up after a deployment is deleted

Signed-off-by: Brian Upton <bupton@vmware.com>