Multi-CPI Director doesn't needlessly recreate VMs
During a `bosh deploy`, a multi-CPI multi-IaaS BOSH Director may
mistakenly recreate VMs that don't need to be recreated; it confuses the
stemcells, thinking that the VM needs a different, wrong stemcell.

This commit fixes that error by fine-tuning the stemcell selection
process. Specifically, the BOSH Director uses the CPI required by the AZ
that the VM is to be placed in to determine which stemcell is desired.
We introduce a new method,
`DeploymentPlan::InstancePlan#stemcell_model_for_cpi`, to properly
select the stemcell.

The source of this stemcell mismatch is from
`DeploymentPlan::Stemcell#add_stemcell_models`, where the first stemcell
is arbitrarily chosen from an array of stemcells, which may be the
correct stemcell for some of the VMs, and the wrong stemcell for others.

Although the decision to make the new method `stemcell_model_for_cpi`
public is debatable (it should be a private method; it's never called
outside the class), it certainly allowed for easier & more thorough
testing.

Fixes, from `bosh task 14721 --debug`:

> D, [2022-03-08T10:29:59.075291 #14721] [] DEBUG -- DirectorJobRunner: Need to update instance 'vpn-probe/e39b63ac-fb33-4e71-bbea-61eb19b5650f (0)', changes: "stemcell" stemcell_changed? changed FROM: bosh-vsphere-esxi-ubuntu-bionic-go_agent TO: bosh-openstack-kvm-ubuntu-bionic-go_agent on instance vpn-probe/e39b63ac-fb33-4e71-bbea-61eb19b5650f (0)

Drive-by: `#stemcell_changed?`: switched `@existing_instance` to
`existing_instance`; if you've gone to the trouble of setting an
`attr_accessor`, then you may as well take advantage it.

[Fixes #2363]