Remove local copies of bosh releases from the main-bosh-docker image
These are compiled releases and cause problems when bosh-deployment is
trying to upgrade to a new stemcell line. The bosh-deployment pipeline
uses the docker bosh to test bosh-deployment, and if the stemcell in bosh-deployment
does not match the compiled releases copied into the docker image here,
the deploy of the director can fail.

Since we build the docker image with the master branch of bosh-deployment,
this leads to a circular dependency where we can't update the docker file
with the bosh-deployment we need until bosh-deployment goes green, which
it can't do until the docker image is updated.

Signed-off-by: Joseph Palermo <jpalermo@pivotal.io>