add conditional use of ops file for bionic
the current task that runs this already uses STEMCELL_OS as an env var.
we have a job to run brats with bionic but while it currently gets an
bionic release it still deploys a jammy stemcell.

see:
```
https://github.com/cloudfoundry/bosh/blob/4cf53f1ef412db4068c21e13b81059661d6c1986/ci/pipeline.yml#L354-L364
```

and
```
https://github.com/cloudfoundry/bosh/blob/4cf53f1ef412db4068c21e13b81059661d6c1986/ci/tasks/test-brats.yml#L39
```

and
```
https://github.com/cloudfoundry/bosh/blob/4cf53f1ef412db4068c21e13b81059661d6c1986/ci/pipeline.yml#L405-L425
```

this updates the director manifest for the docker director conditionally
to use the right stemcell if the task is run with

```
STEMCELL_OS==ubuntu-bionic
```