Change the ruby-install invocations in the docker files to specify just 3.1
This will keep them closer to the actual patch version the director is using.

We removed the explicit SHA checks because this is something ruby-install actually does
on its own.

The only possible value from specifying the SHA is if we did a manual verification
of that particular version of ruby, which we never do. Previously we would go to the ruby-install
source repos, find the SHA, and copy into the install scripts.

[#182075817]

Signed-off-by: Joseph Palermo <jpalermo@pivotal.io>