Update bump postgres to use the http-resource concourse resource
The git repo for concourse felt a bit odd to us. The tarballs associated
with tags there did not match the contents of the postgres packaged
source tarballs. This commit changes the bump postgres package job to
use the tarballs from the postgres download page.

To accomplish this we had to add a second resource, one for postgres 10
and another for postgres 13.

[#182288132]