Fix for the race condition related to NATs users
Every time that a new VM is added by BOSH the authentication file for NATs
receives a new user so the BOSH Agent of that VM is able to authenticate to NATs.
Changes in this file require a NATs reload. This reload was being made by monit,
but it was proving to be unreliable specifically at director startup. This
change is to remove the responsibility of restarting NATs from monit and move it
to the bosh-nats-sync program, which is more flexible since we control its
behavior.

NATs reload will happen in the following scenarios:
- NATs is up and running and bosh-nats-sync is starting.
- NATs is up and running and bosh-nats-sync finds a change in the bosh instances
  that are running.

If for some reason bosh-nats-sync is unable to reload NATs, it will raise an
error and monit will restart it.

For the bosh-nats-sync to be able to reload NATs, privileges were given to its
container, so it's able to send signals to the NATs server process.

Signed-off-by: Daniel Felipe Ochoa <danielfelipo@vmware.com>
Signed-off-by: Brian Upton <bupton@vmware.com>