Updates postgres version to 10.22
For some reason, AWS RDS decided to upgrade our Postgres DB from 10.19
to 10.21. We don't _know_ why, but have found a CVE (CVE-2022-1552) that
was fixed in 10.21, which makes us wonder if Amazon decided to
"helpfully" upgrade our DB to 10.21, and also remove all access to
10.19. Anyway, because of this, we have decided to use the latest 10.x,
which is 10.22.

For reference, this is the complaint that Terraform made that alerted us
to this issue:
 Error: updating RDS DB Instance (terraform-20220610151549614200000002): InvalidParameterCombination: Cannot upgrade postgres from 10.21 to 10.19

And the job in which it happened:
<https://ci.bosh-ecosystem.cf-app.com/teams/main/pipelines/bosh-director/jobs/brats-ubuntu-jammy/builds/67>

Signed-off-by: Rajan Agaskar <ragaskar@vmware.com>