Auto-bump verify-multidigest CLI
This commit adds a resource to check the verify-multidigest S3 bucket
for new versions, and a task which replaces the existing blob if a newer
cli exists.

In addition a new `--load-vars-from` / `-l` stanza was added to the
`ci/config.sh` script in order to interpolate the necessary credentials
for the verify-multidigest S3 bucket.