Remove `jq` installation from db cleanup script
This is causing failures because the apt sources list on the xenial
based `bosh/main`, based on `bosh/main-ruby-go` has outdated links for
Postgres packages and the `apt-get update` command is failing.

Currently the `bosh/main` image contains `jq` so this installation is
unnecessary.

Long term there is cleanup to be done of the images built by and used in
BOSH CI.