Writes basic NATs creds file if UAA is not up
If for some reason NATs sync cannot authenticate to use the BOSH API,
then the user authentication file for NATs will be empty. This causes
problems when UAA is slow to start in some installations.

Our approach to fix this issue was to make NATS Sync write the
credentials for the Director and the Monitor even if the BOSH API is not
available to be consumed. We also made sure that NATs Sync is not going
to overwrite any credentials if they were already written.

[#183572544] Investigate possible NATS 2.0 race condition

Signed-off-by: Daniel Felipe Ochoa <danielfelipo@vmware.com>