do not add more than one scan_and_fix task to the director queue
When a director manages a large number of single instance deployments
in one AZ

And that AZ goes down

Then queuing additional scan_and_fix tasks for a deployment can
create a very large queue for the director containing duplicate
scan_and_fix tasks for the same deployments.

If we already have a scan and fix task queued or being processed
for a particular deployment, there is no additional benefit in
queuing more tasks.

This PR adds a check and some tests around that scenario to avoid
overloading a director that is currently dealing with a large number of
deployments in a bad state.

[ #183759912 ]