add authentication header to get request.
the current sync_client was only used to access the info endpoint. the
info endpoint is unathenticated. Checking the queued tasks for a
deployment requires authentication. Add optional auth header to sending
get requests.

unfortunately the sync client seems to not work support the same
convenience of setting the `authorization` header to an array of values.
since we could either see a Bearer Token from UAA or the basic auth
creds array from `auth_provider(director_info).auth_header` this needs
to manually build the auth header for the sync client if we find basic
auth creds array is returned.

this also updates the stub_request for tasks to actually expect an
authorization header to be set