Configure Blubraket to ignore non-inclusive language in vendor dirs
LFX is scanning CFF repos for non-inclusive language, however a lot of
the findings are for vendored dependencies which we don't
control. Scans are performed using Blubraket which support igore
files:

https://docs.blubracket.com/how-to/cli/using-cli-with-git/#blubracket-ignore

This PR sets up such a file to try and reduce the noice generated by
the LFX scans.

The resuling dashboard can be found here:

https://security.lfx.linuxfoundation.org/#/a092M00001IbM3gQAF/non-inclusive-language