make task state db calls more resilient against flaky pg
in some environments postgres may be overloaded or not working properly.
This can lead to connections being unexpectedly closed or failing. In a
scenario where a `delayed job` execuction finished (and the job was
updated by the `delayed job` gem in the `delayed job table` the update
for the task itself in the `tasks` table may fail. The worker will not
pick up the task again (as the data in `delayed jobs` table indicates it
finished, but the tasks table (which bosh uses to display `bosh tasks`)
leaves the task in the `queued` or another state.

Fixes:
`Bosh::Director::Jobs::DBJob (id=897735) (queue=urgent) FAILED (0 prior attempts) with Sequel::DatabaseDisconnectError: PG::UnableToSend: server closed the connection unexpectedly`

Currently some of the task update logic was already wrapped in a
repeatable transaction but the errors we've seen in the user-report
indicated that updating the task fail with the above error.

While we cannot ensure that the task state will be updated in 100% of
the cases, this PR  minimizes the chance that the task state will not be
updated when the connection to postgres is flaky.

This behaviour can be reproduced by executing the below sql statement
against the PG instance backing the director

```
SELECT
 pg_terminate_backend(pid)
FROM
 pg_stat_activity
WHERE
 -- don't kill my own connection!
 pid <> pg_backend_pid()
 -- don't kill the connections to other databases
 AND datname = 'bosh';
pg_terminate_backend
```

while running bosh arbitrary bosh tasks.