mark orphaned tasks that are not backed by a delayed job as errored
in some cases when the worker fails to update the task state when a
delayed job exits (either with error or success) we can have orphaned
tasks polluting the tasks db. Since the delayed job execution already
finished and it failed updating the task state (but not the delayed_jobs
table), there is nothing that will "finalize" the task. It will be "stuck"
in it's last state before the delayed job process exited.

To deal with this we can look at the current items in delayed_jobs
table. only tasks that still have a chance of getting their state
updated will have an entry here (until the delayed_job worker finishes
running it). We explicitly tell delayed_jobs to delete errord finished
jobs too.

It should be safe to mark all tasks that are in queued / processing
state as errored when we find no entry in the delayed_jobs db which has
a handler for our task_id