merge integration and ruby-go-main image
the only difference is that integration also contains java runtime.

we're trying to have less images throughout to improve on maintenance
efforts.

make all images based on the integration image as well as trigger their
builds.

Update remaining images to be based on integration.

the db-type images were already based on main-ruby-go and just added a
DB and some config for the DB.

docker-cpi image purpose is to add docker and some setup to it can be used to start
a priviledged container that deploys a bosh with a docker cpi.