remove transaction and retry
this change was added to try to make the DB calls here more resilient.
As pointed out in the review this transaction is not idempotent and may
lead to issues down the road. Since this area of the code wasn't involved
in causing the stuck tasks this part of d59e3a76decb6055310e903b09e8854669493dbd
is getting removed again.