address idempotency of transaction
a review pointed out that a transaction that fails on the commit itself
( commit received and executed by PG ) by losing the connection and
the client never learning about the success of the commit would lead to
a situaiton where task.state is already updated and would be updated
again.

moving the initial task read from the DB outside the transaction and the
caching the value of task.state should adress that. Even if the
transaction fails as described, the 2nd execution would still use the
cached value for the switch and therefore write the right value into
the db.