feat: Short Lived NATS credentials
When bootstraping a new VM, BOSH stores credentials in the metadata
services of the different IAASes. This includes the certificates the VM
uses to access NATS. This is a security concern for some environments
where a malicious user that can access the metadata service gets the
NATS credentials. He will be able to connect to NATS and listen to the
traffic between the the VMs and the director.

This feature removes that risk by making the credentials stored in
the metadata service ephemeral/short lived, so if a user gets access
to them, they won't be useful after the VM bootstrap. These credentials
will be used only while starting a new VM, and after that, they will be
replaced by permanent credentials that are not stored in the metadata
service.

This feature is enabled by default. To disable it a new parameter
called `enable_short_lived_nats_credentials` should be added with false as
its value in the director manifest during the create-env.
This feature will affect only newly created/recreated VMs.

When an installation is using this feature its a MUST to ensure that the
stemcells used in the director are compatible with it; stemcells that
are incompatible will result in unresponsive VMs.
The oldest compatible stemcell versions are:
Windows 2019 - 2019.41
Ubuntu Xenial - 621.171
Ubuntu Bionic - 1.36
Ubuntu Jammy - All versions are suported.

[#183316690] [Deadline: Before End of FYQ1] [AHA] Short-lived NATS bootstrap credential

Co-authored-by: Daniel Felipe Ochoa <danielfelipo@vmware.com>
Co-authored-by: Brian Upton <bupton@vmware.com>
Co-authored-by: Brian Cunnie <bcunnie@vmware.com>