Add flag: ephemeral nats creds on compilation VMs
NATS credentials could short lived in all the VMs, but for some special
scenarios we want more control so the rotation could be
disabled/enabled. Compilation VMs is a special case and for it we want
to default to not rotate the NATs credentials.

In this commit we introduce the flag
`enable_short_lived_nats_bootstrap_credentials_compilation_vms` and its
default value is false. This will prevent BOSH from rotating NATS
credentials on these type of VMs.

[#183316690] [Deadline: Before End of FYQ1] [AHA] Short-lived NATS bootstrap credential