Disable short lived nats creds on the sandbox
Testing on the sandbox is limited and lacks a manager for the bosh-agent
process. This makes imposible for the agent to restart after it is
killed. Because of that we cannot run the short lived nats credentials
feature in the sandbox. This change is to make the feature disabled
while running in that environment.

[#183316690] [Deadline: Before End of FYQ1] [AHA] Short-lived NATS bootstrap credential