Remove Ruby SHA512 checksum
This was introduced recently, as previous iterations of the Dockerfile
installed Ruby without an explicit checksum check. Since we are not
pinning to a specific patch, as new versions of Ruby 3.2 are made
available, the checksum would need to be updated.

For now, we should be able to rely on the internal verification done by
the ruby-install script.

Signed-off-by: Kenneth Lakin <klakin@vmware.com>