Removing tests for stemcells without signed URLs
This test is removed because we're not supporting the Xenial 456 line
anymore and it was designed to ensure compatibility with that line.
Also the new feature of short lived NATS credentials is incompatible
with those old stemcells and it was causing the BRATS to fail.

Since the line is no longer supported it was easier to remove the test
than disabled the short lived NATS credentials for this test.

[#183316690] [Deadline: Before End of FYQ1] [AHA] Short-lived NATS bootstrap credential

Signed-off-by: Daniel Felipe Ochoa <danielfelipo@vmware.com>