Rebuilds Postgres indexes on glibc version change
This commit:
* Adds early fails to the Postgres 10 and Postgres 13 jobs if the
  Postgres 15 data directory is present. This omission seemed to be a
  clear oversight.
* Adds code to detect glibc version changes and rebuild indexes on all
  Postgres databases if the running glibc is newer than 2.27, and has
  changed since the last time we rebuild the indexes.

glibc version 2.28 introduced a major change to how it performs
collation. Whenever any changes to collations are introduced, indexes on
Postgres tables must be rebuilt in order to avoid problems which include
successful insertion of rows that violate uniqueness constraints on some
keys but not others.

Postgres 15 and later includes code that detects and warns about this
situation, but leaves correcting the issue up to the operator... so the
updates in this commit are still required.

Signed-off-by: Joseph Palermo <jpalermo@pivotal.io>