Add new brats package to measure template rendering performance.
Originally I created a fake bosh release to test against, but getting the performance
profile to match real world problems was problematic. So using cf-deployment directly
makes more sense as that is a place we have seen real world problems.

These are not intended to have assertions or other steps of the build use them as a
passed constraint. They are informational only so we can see how changes impact template
rendering performance.