Sets internal Postgres shared_buffers to default
The default for this value is 128MB, and has been since at least
Postgresql 9.3. This isn't causing any _obvious_ problems, but:

a) The current value is 1/4th of the default value.
b) An additional 96MB of RAM is not much, given that the _default_ size
   of a Director VM is 8GB.
c) This setting was changed a _bunch_ throughout the history of the Bosh
   repo. The choice of value was never justified, and the change was
   almost always part of a much larger, and entirely unrelated set of
   changes. So, we feel that going with the defaults recommended by the
   Postgres developers is a much better choice than some
   arbitrarily-chosen value from the distant past.
d) Maybe this resolves the somewhat-frequent "Your postgres connection
   was disconnected unexpectedly" messages that we often see in the logs
   of healthy Directors, who knows?

Signed-off-by: Daniel Felipe Ochoa <danielfelipo@vmware.com>