Fix for "Extracting compiled package archive failed" bug when uploading release with compiled packages
This situation exists if the bosh director already has a compiled package with a fingerprint and the
cli is trying to upload a release that has two different copies of that same package with the same
fingerprint using two different names. The current package matching logic only uses fingerprints to
match, so the cli ends up removing both packages from the release before uploading it, but the director
doesn't actually have compiled versions of both packages.

So instead, we now do not send the fingerprint in the matching response if there is a package
in the release manifest that includes that fingerprint but was not successfully matched.