Allow `max_in_flight_overrides` for problem resolution
The put '/:deployment/problems' endpoint now supports a
`max_in_flight_overrides` key in the request body that is a map from
instance group name to `max_in_flight` value.  If an override is
provided for a given instance group, it will be passed through to the
`ProblemResolver` and used instead of the `max_in_flight` value from the
manifest.  This will allow for the potential of a faster disaster
recovery.

[#185547787]

Authored-by: Chris Selzo <cselzo@vmware.com>