Add CI task to check for updated blobs
The check-for-updated-blobs task compares config/blobs.yml with the last
released version to see if a given list of blobs has been updated. It
follows a similar pattern to the golang-release tasks, exiting non-zero
when there are no updates so that you can use `on_success` to bump the
version automatically.

[#185631201] Ensure components using Golang are auto-released (as well as auto-bumped)