Stop using pre-release versions
The pipeline currently adds an `rc` pre-release version to the semver
resource everytime a commit passes the candidate-release job. This
practice needs to change because the auto-release CI pattern requires
triggering off of new versions; so creating pre-release versions would
unintentionally trigger new releases.

This change should essentially be a no-op because the pre-release
version resource was only ever consumed by the `finalize-release` job,
which immediately did a `bump: final` to remove the pre-release portion.
Additionally, the exist `bump-major/minor/patch` jobs, which are
required to be run before `finalize-release` also caused the pre-release
portion to be dropped.

[#185631201] Ensure components using Golang are auto-released (as well as auto-bumped)