Updates nginx config to use supported syntax
This commit finishes the work started in Story #168153036 to update the
syntax of the Director's nginx configuration files to remove the
deprecated 'ssl' server directive and move to the supported
'listen <PORT> ssl' server directive.

On Thursday, August 24th we pushed a new version of
bosh-package-nginx-release (which -oddly- is not on bosh.io). This moved
us from nginx 1.21.4 to nginx 1.25.2. According to the nginx changelog,
support for the 'ssl' server directive was finally removed. Unlucky us.

[#168153036] Update nginx conf to remove deprecations/warnings