Explicitly provide ActiveSupport extensions in ERB
We previously upgraded from `activesupport` 7.0.8 to 7.1.0. This broke
some releases that were using the `Object#present?` method provided by
ActiveSupport. It seems that the extensions it provides are no longer
auto-included. We were getting these methods accidentally because the
BOSH Director uses both delayed_job and factory_bot, which both
transitively include activesupport.

Because releases in the wild are relying on these methods, we now
explicitly include the `active_support/core_ext` library in order to
continue providing the methods.

[#186295738] BOSH template rendering fails when `#present?` is used

Signed-off-by: Chris Selzo <cselzo@vmware.com>