Stop unnecessarily scan the 'job.MF' file from job blob when rendering ERBs
Indeed we already know the job name and template list, as these data are
persisted when a release is uploaded to the director.

This makes unnecessary to verify that the job name is consistent at ERB
rendering time.

And this consistency check is more relevant when done at the time a release
job is uploaded to the director.