Populate stemcell os when manifest contains only the name
Follow up to: https://github.com/cloudfoundry/bosh/pull/2475

The previous PR didn't handle diffs correctly so it showed that runtime configs
were getting added/removed even though they weren't.

This commit moves the OS population up into the manifest parsing so it can apply
to both the deployment manifest and the diff manifest.

This caused the DeploymentPlan Stemcell to start failing because it had a validation
that both Name and OS shouldn't be specified. We moved that validation up to the
ManifestValidator which seems like a more appropriate place for it anyway.

We modified the manifest parsing to prefer Name over OS when resolving the stemcell version
aliases. This is more consistent with how the DeploymentPlan Stemcell resolves Name and OS.
It shouldn't be a behavioral change though since you can't specify both Name and OS.

Signed-off-by: Joseph Palermo <joseph.palermo@broadcom.com>