Follow up to #2484, which broke everything...
Although those changes worked for full deploys, the way that manifests
get parsed when doing things such as recreates causes the jobs to trip over the
"cannot specify both name and os for stemcell" validation.

It would probably be possible to find some way to work around this, but the
validation just isn't very useful anyway.

This commit removes that validation, making it valid now to specify both name and os.

When both are specified, name is given priority and os is ignored (as name has the os included
in it anyway). It is now possible to make an invalid manifest, where the name and os don't
actually match, but it will deploy anyway as long as the name is valid. This doesn't
seem like a compelling reason to keep this validation around.